+++ e2e/mocks/microservice.controller.ts
@@ -1,5 +1,5 @@
 import { Controller } from '@nestjs/common';
-import { RMQMessage, RMQError, RMQRoute, Validate, ExtendedMessage, RMQService } from '../../lib';
+import { RMQMessage, RMQError, RMQRoute, RMQValidate, ExtendedMessage, RMQService, RMQTransform } from '../../lib';
 import {
 	DivideContracts,
 	MultiplyContracts,
@@ -8,7 +8,7 @@ import {
 	TimeOutContracts,
 	AppIdContracts,
 	ManualAckContracts,
-	DebugContracts, CustomMessageFactoryContracts, PatternStarContracts, PatternHashContracts,
+	DebugContracts, CustomMessageFactoryContracts, PatternStarContracts, PatternHashContracts, TransformContracts,
 } from '../contracts/mock.contracts';
 import { ERROR_TYPE } from '../../lib/constants';
 import { Message } from 'amqplib';
@@ -18,7 +18,7 @@ export class MicroserviceController {
 	constructor(private readonly rmqService: RMQService) { }
 
 	@RMQRoute(SumContracts.topic)
-	@Validate()
+	@RMQValidate()
 	sumRpc({ arrayToSum }: SumContracts.Request): SumContracts.Response {
 		const result = arrayToSum.reduce((prev, cur) => prev + cur);
 		if (result === 0) {
@@ -34,20 +34,20 @@ export class MicroserviceController {
 	}
 
 	@RMQRoute(NotificationContracts.topic)
-	@Validate()
+	@RMQValidate()
 	notificationNone({ message }: NotificationContracts.Request): void {
 		console.log(message);
 		return;
 	}
 
 	@RMQRoute(MultiplyContracts.topic)
-	@Validate()
+	@RMQValidate()
 	multiplyRpc({ arrayToMultiply }: MultiplyContracts.Request): MultiplyContracts.Response {
 		return { result: arrayToMultiply.reduce((prev, cur) => prev * cur) };
 	}
 
 	@RMQRoute(DivideContracts.topic)
-	@Validate()
+	@RMQValidate()
 	divide({ first, second }: DivideContracts.Request): DivideContracts.Response {
 		return { result: first / second };
 	}
@@ -98,5 +98,11 @@ export class MicroserviceController {
 	@RMQRoute(PatternHashContracts.topic)
 	hashPattern({ num }: PatternHashContracts.Request): PatternHashContracts.Response {
 		return { num };
+	}
+
+	@RMQTransform()
+	@RMQRoute(TransformContracts.topic)
+	transform({ date }: TransformContracts.Request): TransformContracts.Response {
+		return { res: date.getFullYear() + 1, type: typeof date };
 	}
 }