+++ lib/interfaces/rmq-service.interface.ts
@@ -5,8 +5,8 @@ export interface IRMQService {
 	init: () => Promise<void>;
 	ack: (...params: Parameters<Channel['ack']>) => ReturnType<Channel['ack']>;
 	nack: (...params: Parameters<Channel['nack']>) => ReturnType<Channel['nack']>;
-	send: <IMessage, IReply>(topic: string, message: IMessage, options?: IPublishOptions) => Promise<IReply>
-	notify: <IMessage>(topic: string, message: IMessage, options?: IPublishOptions) => Promise<void>
+	send: <IMessage, IReply>(topic: string, message: IMessage, options?: IPublishOptions) => Promise<IReply>;
+	notify: <IMessage>(topic: string, message: IMessage, options?: IPublishOptions) => Promise<void>;
 	healthCheck: () => boolean;
 	disconnect: () => Promise<void>;
 	mockReply?: <T>(topic: string, reply: T) => void;